Switch default handling of py floats as RS floats (#548)
* switch default handling of py floats as rs floats

* update tests for new default float data type

* change constant name to FLOAT

+++ parsons/databases/redshift/constants.py
@@ -1,6 +1,6 @@
 COL_NAME_MAX_LEN = 120
 
-DECIMAL = 'decimal'
+FLOAT = 'float'
 
 REPLACE_CHARS = {" ": ""}
